+++ packages/react-storage/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @aws-amplify/ui-react-storage
 
+## 3.0.8
+
+### Patch Changes
+
+- Updated dependencies [[`4f643b05b`](https://github.com/aws-amplify/amplify-ui/commit/4f643b05b010901226bf200f8d0b06601d0ecab5), [`190bf28d1`](https://github.com/aws-amplify/amplify-ui/commit/190bf28d1b570329f415d2d4c03e608a3b30412b)]:
+  - @aws-amplify/ui-react-core@3.0.6
+  - @aws-amplify/ui@6.0.6
+  - @aws-amplify/ui-react@6.0.7
+
 ## 3.0.7
 
 ### Patch Changes